Move descriptors into registry.py from util.py to avoid circular import
