+++ skbio/io/util.py
@@ -37,6 +37,8 @@ from skbio.io._fileobject import (
     CompressedBufferedReader,
     CompressedBufferedWriter,
 )
+
+# from skbio.io.registry import io_registry
 import skbio
 
 _d = dict(
@@ -298,85 +300,88 @@ def open_files(files, **kwargs):
         yield [stack.enter_context(open_file(f, **kwargs)) for f in files]
 
 
-class Read:
-    """A descriptor class to generate read methods for scikit-bio objects."""
+# class Read:
+#     """A descriptor class to generate read methods for scikit-bio objects."""
 
-    def __get__(self, instance, cls):
-        return self._generate_read_method(cls)
+#     def __get__(self, instance, cls):
+#         return self._generate_read_method(cls)
 
-    def _generate_read_method(self, cls):
-        def read_method(file, format=None, **kwargs):
-            return skbio.io.read(file, into=cls, format=format, **kwargs)
+#     def _generate_read_method(self, cls):
+#         def read_method(file, format=None, **kwargs):
+#             return skbio.io.read(file, into=cls, format=format, **kwargs)
 
-        read_method.__doc__ = self._make_docstring(cls)
+#         read_method.__doc__ = self._make_docstring(cls)
 
-        return read_method
+#         return read_method
 
-    def _make_docstring(self, cls):
-        return f"""Create a new ``{cls.__name__}`` instance from a file.
+#     def _make_docstring(self, cls):
+#         supported_formats = io_registry.list_read_formats(cls)
+#         return f"""Create a new ``{cls.__name__}`` instance from a file.
 
-This is a convenience method for :func:`skbio.io.read`. For more information about the
-I/O system in scikit-bio, please see :mod:`skbio.io`.
+# This is a convenience method for :func:`skbio.io.registry.read`. For more information
+# about the I/O system in scikit-bio, please see :mod:`skbio.io`.
 
-Supported file formats include:
+# Supported file formats include:
 
-something like {cls.__name__}.list_read_formats
+# {supported_formats}
 
-Parameters
-----------
-file : openable (filepath, URL, filehandle, etc.)
-    The location to read the given `format` into. Something that is understood by
-    :func:`skbio.io.util.open`. Filehandles are not automatically closed, it is the
-    responsibility of the caller.
-format : str, optional
-    The format of the file. If None, the format will be inferred.
-kwargs : dict, optional
-    Additional arguments passed to the reader for {cls.__name__}.
+# Parameters
+# ----------
+# file : openable (filepath, URL, filehandle, etc.)
+#     The location to read the given `format` into. Something that is understood by
+#     :func:`skbio.io.util.open`. Filehandles are not automatically closed, it is the
+#     responsibility of the caller.
+# format : str, optional
+#     The format of the file. The format must be a format name with a reader for
+#     {cls.__name}. If None, the format will be inferred.
+# kwargs : dict, optional
+#     Additional arguments passed to :func:`skbio.io.registry.read()` and the reader for
+#     {cls.__name__}.
 
-Returns
--------
-{cls.__name__}
-    A new instance of {cls.__name__}.
-"""
+# Returns
+# -------
+# {cls.__name__}
+#     A new instance of {cls.__name__}.
+# """
 
 
-class Write:
-    """A descriptor class to generate write methods for scikit-bio objects."""
+# class Write:
+#     """A descriptor class to generate write methods for scikit-bio objects."""
 
-    def __get__(self, instance, cls):
-        if instance is None:
-            return self._generate_write_method(cls)
-        else:
-            return MethodType(self._generate_write_method(cls), instance)
+#     def __get__(self, instance, cls):
+#         if instance is None:
+#             return self._generate_write_method(cls)
+#         else:
+#             return MethodType(self._generate_write_method(cls), instance)
 
-    def _generate_write_method(self, cls):
-        def write_method(self, file, format=None, **kwargs):
-            if format is None:
-                if hasattr(cls, "default_write_format"):
-                    format = cls.default_write_format
-                else:
-                    raise ValueError(f"{cls.__name__} has no default write format.")
-            return skbio.io.write(self, into=file, format=format, **kwargs)
+#     def _generate_write_method(self, cls):
+#         def write_method(self, file, format=None, **kwargs):
+#             if format is None:
+#                 if hasattr(cls, "default_write_format"):
+#                     format = cls.default_write_format
+#                 else:
+#                     raise ValueError(f"{cls.__name__} has no default write format.")
+#             return skbio.io.write(self, into=file, format=format, **kwargs)
 
-        write_method.__doc__ = self._make_docstring(cls)
+#         write_method.__doc__ = self._make_docstring(cls)
 
-        return write_method
+#         return write_method
 
-    def _make_docstring(self, cls):
-        return f"""Write the {cls.__name__} instance to a file.
+#     def _make_docstring(self, cls):
+#         return f"""Write the {cls.__name__} instance to a file.
 
-This is a convenience method for :func:`skbio.io.write`.
+# This is a convenience method for :func:`skbio.io.write`.
 
-Parameters
-----------
-file : openable (filepath, URL, filehandle, etc.)
-    The location to write the given `format`.
-format : str, optional
-    The format to write the {cls.__name__} object as.
-kwargs : dict, optional
-    Additional arguments passed to the writer for {cls.__name__}.
+# Parameters
+# ----------
+# file : openable (filepath, URL, filehandle, etc.)
+#     The location to write the given `format`.
+# format : str, optional
+#     The format to write the {cls.__name__} object as.
+# kwargs : dict, optional
+#     Additional arguments passed to the writer for {cls.__name__}.
 
-Returns
--------
-None
-"""
+# Returns
+# -------
+# None
+# """
