+++ skbio/sequence/_rna.py
@@ -10,7 +10,7 @@ import skbio
 from skbio.util._decorator import classproperty, overrides
 from ._nucleotide_mixin import NucleotideMixin, _motifs as _parent_motifs
 from ._grammared_sequence import GrammaredSequence
-from skbio.io.util import Read, Write
+from skbio.io.registry import Read, Write
 
 
 class RNA(GrammaredSequence, NucleotideMixin):
