+++ skbio/tree/_tree.py
@@ -29,7 +29,7 @@ from skbio.tree._exception import (
 from skbio.util import get_rng, RepresentationWarning
 from skbio.util._decorator import classonlymethod
 from skbio.util._warning import _warn_deprecated
-from skbio.io.util import Read, Write
+from skbio.io.registry import Read, Write
 
 
 def distance_from_r(m1, m2):
